move QModule and VModule to layers
