+++ all/approximation/q_network.py
@@ -1,27 +1,11 @@
-import torch
-from torch import nn
 from torch.nn.functional import mse_loss
-from all.layers import ListNetwork
+from all.layers import QModule
 from .approximation import Approximation
 
 def td_loss(loss):
     def _loss(estimates, errors):
         return loss(estimates, errors + estimates.detach())
     return _loss
-
-class QModule(nn.Module):
-    def __init__(self, model, num_actions):
-        super().__init__()
-        self.device = next(model.parameters()).device
-        self.model = ListNetwork(model, (num_actions,))
-
-    def forward(self, states, actions=None):
-        values = self.model(states)
-        if actions is None:
-            return values
-        if isinstance(actions, list):
-            actions = torch.tensor(actions, device=self.device)
-        return values.gather(1, actions.view(-1, 1)).squeeze(1)
 
 class QNetwork(Approximation):
     def __init__(
