+++ all/approximation/v_network.py
@@ -1,21 +1,11 @@
-from torch import nn
 from torch.nn.functional import mse_loss
-from all.layers import ListNetwork
+from all.layers import VModule
 from .approximation import Approximation
 
 def td_loss(loss):
     def _loss(estimates, errors):
         return loss(estimates, errors + estimates.detach())
     return _loss
-
-class VModule(nn.Module):
-    def __init__(self, model):
-        super().__init__()
-        self.device = next(model.parameters()).device
-        self.model = ListNetwork(model, (1,))
-
-    def forward(self, states):
-        return self.model(states).squeeze(-1)
 
 class ValueNetwork(Approximation):
     def __init__(
