+++ all/layers/__init__.py
@@ -120,5 +120,25 @@ class Linear0(nn.Linear):
         if self.bias is not None:
             nn.init.constant_(self.bias, 0.)
 
+class QModule(nn.Module):
+    def __init__(self, model, num_actions):
+        super().__init__()
+        self.device = next(model.parameters()).device
+        self.model = ListNetwork(model, (num_actions,))
 
-__all__ = ["Aggregation", "Dueling", "Flatten", "NoisyLinear", "Linear0"]
+    def forward(self, states, actions=None):
+        values = self.model(states)
+        if actions is None:
+            return values
+        if isinstance(actions, list):
+            actions = torch.tensor(actions, device=self.device)
+        return values.gather(1, actions.view(-1, 1)).squeeze(1)
+
+class VModule(nn.Module):
+    def __init__(self, model):
+        super().__init__()
+        self.device = next(model.parameters()).device
+        self.model = ListNetwork(model, (1,))
+
+    def forward(self, states):
+        return self.model(states).squeeze(-1)
